已成功修复 GCC 编译器构建产生的所有项目代码警告。
修复内容
类型转换警告 - 添加显式 static_cast 转换
Switch 语句警告 - 补全 ErrorCode 枚举处理
未使用代码警告 - 删除未使用函数/变量
ODR 违规警告 - 重命名 ReorderMapData 避免冲突
Clang 特有警告 - 构造函数参数遮蔽修复
构建验证
✅ GCC Release 构建成功，无项目代码警告
⚠️ Clang Release 代码警告已修复，链接器配置问题需单独处理
备注
仅剩 quill 第三方库的 LTO 警告（库内部问题无法修复）
变更日志已保存至 changelog/2026-01-29-compiler-warnings-fix.md

+++ src/algo/pe_optimizer.cpp
@@ -156,8 +156,8 @@ std::pair<bool, std::size_t> PEOptimizer::checkComplementarity(
     }
 
     // Add length difference
-    diffCount += std::abs(static_cast<int>(r1Seq.length()) -
-                          static_cast<int>(r2Seq.length()));
+    diffCount += static_cast<std::size_t>(std::abs(static_cast<int>(r1Seq.length()) -
+                          static_cast<int>(r2Seq.length())));
 
     bool beneficial = diffCount <= config_.complementarityThreshold;
     return {beneficial, diffCount};
